Revert "Fix default icon shown for BGMApp bundle."
This reverts commit a4160d370d2c447d0f44c80548a1d0648b82eefe.

I've tested the icon bug on a couple of other machines now and haven't been
able to reproduce it. My "fix" did cause the bug on those machines, though,
despite fixing it on mine. Still not sure exactly what's going on, but it seems
to be a problem with my development environment.