Fix another crash when BGMDevice's volume is changed.
Fixes the same issue as df9815a4be622a3e86ad00e23a18b13152a22937, but in
BGMOutputVolumeMenuItem instead of BGMStatusBarItem.

I think the problem was that it captured a weak reference in a C++
lambda, but it would capture by (C++) reference and when the reference
was used it would be referencing invalid memory. The fix to have the
lambda capture by value instead.

See #202.