Try to avoid using the discrete GPU on multi-GPU systems to save power.
Add NSSupportsAutomaticGraphicsSwitching to BGMApp's Info.plist in the
hopes that it will use the integrated GPU when possible. BGMApp doesn't
need to use the fastest GPU, so it's better to save battery power.

Untested because I don't have the right hardware. (I actually have an
iMac with a discrete GPU, but the integrated GPU seems to be disabled.)

See #284 and
<https://developer.apple.com/library/archive/qa/qa1734/_index.html>.